Migrate project structure to getgb.io
