+++ .travis.yml
@@ -2,10 +2,10 @@ language: go
 go:
   - 1.5
   - 1.6
+install:
+  - go get github.com/constabulary/gb/...
+  - ./vendor.sh
 script:
-  - curl -s https://raw.githubusercontent.com/pote/gpm/v1.3.2/bin/gpm > gpm
-  - chmod +x gpm
-  - ./gpm install
   - ./test.sh
 notifications:
   email: false
