--- Godeps
@@ -1,2 +1,0 @@
-github.com/riobard/go-mmap              8eec19e37d25c6568d153517fa0214fbae68a2f1
-github.com/bitly/timer_metrics          afad1794bb13e2a094720aeb27c088aa64564895
