+++ README.md
@@ -19,14 +19,6 @@ Pre-built binaries for linux and darwin are available for download:
 * [sortdb-1.0.darwin-amd64.go1.4.2.tar.gz](https://github.com/jehiah/sortdb/releases/download/v1.0/sortdb-1.0.darwin-amd64.go1.4.2.tar.gz)
 * [sortdb-1.0.linux-amd64.go1.4.2.tar.gz](https://github.com/jehiah/sortdb/releases/download/v1.0/sortdb-1.0.linux-amd64.go1.4.2.tar.gz)
 
-**Building From Source**
-
-Pre-requisites:
-
-* golang (version 1.2+ is required)
-* gpm (dependency manager)
-
-`sortdb` uses gpm to manage dependencies and produce reliable builds. Using gpm is the preferred method when compiling from source.
 
 ### Usage
 
@@ -110,6 +102,12 @@ LC_COLLATE=C sort data.csv > sorted_data.csv
 ```
 
 Note: The locale specified by the environment affects sort order. Set `LC_ALL=C` or `LC_COLLATE=C` to get the traditional sort order that uses native byte values.
+
+## Building From Source
+
+This project uses [gb](https://getgb.io/) the Go Build tool. 
+
+To load vendored dependencies and compile run `./vendor.sh && gb build`
 
 --
