+++ dist.sh
@@ -5,27 +5,31 @@ set -e
 
 DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 echo "working dir $DIR"
-mkdir -p $DIR/dist
-mkdir -p $DIR/.godeps
-export GOPATH="$DIR/.godeps:$GOPATH"
-GOPATH=$DIR/.godeps gpm install
 
-
-os=$(go env GOOS)
-arch=$(go env GOARCH)
-version=$(cat $DIR/version.go | grep "const VERSION" | awk '{print $NF}' | sed 's/"//g')
-goversion=$(go version | awk '{print $3}')
+rm -rf $DIR/vendor
+echo "... refreshing vendor directory"
+./vendor.sh
 
 echo "... running tests"
 ./test.sh || exit 1
+
+arch=$(go env GOARCH)
+version=$(cat $DIR/src/cmd/sortdb/version.go | grep "const VERSION" | awk '{print $NF}' | sed 's/"//g')
+goversion=$(go version | awk '{print $3}')
 
 for os in linux darwin; do
     echo "... building v$version for $os/$arch"
     BUILD=$(mktemp -d -t sortdb)
     TARGET="sortdb-$version.$os-$arch.$goversion"
-    GOOS=$os GOARCH=$arch CGO_ENABLED=0 go build -o $BUILD/$TARGET/sortdb || exit 1
-    pushd $BUILD
+    GOOS=$os GOARCH=$arch CGO_ENABLED=0 gb build
+    mkdir -p $BUILD/$TARGET
+    cp bin/sortdb-$os-$arch $BUILD/$TARGET/sortdb
+    pushd $BUILD >/dev/null
     tar czvf $TARGET.tar.gz $TARGET
+    if [ -e $DIR/dist/$TARGET.tar.gz ]; then
+        echo "... WARNING overwriting dist/$TARGET.tar.gz"
+    fi
     mv $TARGET.tar.gz $DIR/dist
-    popd
+    echo "... built dist/$TARGET.tar.gz"
+    popd >/dev/null
 done
