+++ src/cmd/sortdb/version.go
@@ -1,0 +1,3 @@
+package main
+
+const VERSION = "1.2.0-alpha"
