+++ test.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 set -e
 
-go test -timeout 60s ./...
-GOMAXPROCS=4 go test -timeout 60s -race ./...
+gb test -timeout 60s
+# gb test -timeout 60s -race
+gb build
