--- version.go
@@ -1,3 +1,0 @@
-package main
-
-const VERSION = "1.1"
